+++ backend/src/index.ts
@@ -32,9 +32,9 @@ app.use(cors());
 // Stands in for the body-parser used in BlockLand; parses incoming requests into JSON format and enables you to access the request BODY:
 app.use(express.json());
 // // Set the public directory as the server's static 
-// app.use(express.static("public"));
 
 // Server Endpoints
+app.use(require('./endpoints/loginEndpoints'));
 app.get('/api/:dbName/:collection', getTestData);
 // Test endpoint for communication with the frontend:
 app.get('/api/test', getFrontend);